requirements.txt
astroid==2.11.6
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
dill==0.3.5.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysql-connector-python==8.0.29
platformdirs==2.5.2
protobuf==4.21.2
pycodestyle==2.8.0
pylint==2.14.3
requests==2.28.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
urllib3==1.26.9
wrapt==1.14.1
